% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamtofastq.R
\name{bamtofastq}
\alias{bamtofastq}
\title{Convert BAM alignments to FASTQ files.}
\usage{
bamtofastq(i, fq, fq2 = NULL, tags = NULL)
}
\arguments{
\item{i}{<BAM>}

\item{fq}{<FQ>}

\item{fq2}{FASTQ for second end.  Used if BAM contains paired-end data.
BAM should be sorted by query name is creating paired FASTQ.}

\item{tags}{Create FASTQ based on the mate info
in the BAM R2 and Q2 tags.}
}
\description{
Convert BAM alignments to FASTQ files.
}
