% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.bamtofastq.R
\name{bt.bamtofastq}
\alias{bt.bamtofastq}
\title{Convert BAM alignments to FASTQ files.}
\usage{
bt.bamtofastq(i, fq, fq2 = NULL, tags = NULL, output = NULL)
}
\arguments{
\item{i}{<BAM>}

\item{fq}{<FQ>}

\item{fq2}{FASTQ for second end.  Used if BAM contains paired-end data.
BAM should be sorted by query name is creating paired FASTQ.}

\item{tags}{Create FASTQ based on the mate info
in the BAM R2 and Q2 tags.}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Convert BAM alignments to FASTQ files.
}
