% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.bedpetobam.R
\name{bt.bedpetobam}
\alias{bt.bedpetobam}
\title{Converts feature records to BAM format.}
\usage{
bt.bedpetobam(i, g, mapq = NULL, ubam = NULL, output = NULL)
}
\arguments{
\item{i}{<bed/gff/vcf>}

\item{g}{<genome>}

\item{mapq}{Set the mappinq quality for the BAM records.
(INT) Default: 255}

\item{ubam}{Write uncompressed BAM output. Default writes compressed BAM.}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Converts feature records to BAM format.
}
