% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.bedtobam.R
\name{bt.bedtobam}
\alias{bt.bedtobam}
\title{Converts feature records to BAM format.}
\usage{
bt.bedtobam(i, g, mapq = NULL, bed12 = NULL, ubam = NULL, output = NULL)
}
\arguments{
\item{i}{<bed/gff/vcf>}

\item{g}{<genome>}

\item{mapq}{Set the mappinq quality for the BAM records.
(INT) Default: 255}

\item{bed12}{The BED file is in BED12 format.  The BAM CIGAR
string will reflect BED "blocks".}

\item{ubam}{Write uncompressed BAM output. Default writes compressed BAM.}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Converts feature records to BAM format.
}
