% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.flank.R
\name{bt.flank}
\alias{bt.flank}
\title{Creates flanking interval(s) for each BED/GFF/VCF feature.}
\usage{
bt.flank(
  i,
  g,
  b = NULL,
  l = NULL,
  r = NULL,
  s = NULL,
  pct = NULL,
  header = NULL,
  output = NULL
)
}
\arguments{
\item{i}{<bed/gff/vcf>}

\item{g}{<genome>}

\item{b}{Create flanking interval(s) using -b base pairs in each direction.
- (Integer) or (Float, e.g. 0.1) if used with -pct.}

\item{l}{The number of base pairs that a flank should start from
orig. start coordinate.
- (Integer) or (Float, e.g. 0.1) if used with -pct.}

\item{r}{The number of base pairs that a flank should end from
orig. end coordinate.
- (Integer) or (Float, e.g. 0.1) if used with -pct.}

\item{s}{Define -l and -r based on strand.
E.g. if used, -l 500 for a negative-stranded feature, 
it will start the flank 500 bp downstream.  Default = false.}

\item{pct}{Define -l and -r as a fraction of the feature's length.
E.g. if used on a 1000bp feature, -l 0.50, 
will add 500 bp "upstream".  Default = false.}

\item{header}{Print the header from the input file prior to results.}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Creates flanking interval(s) for each BED/GFF/VCF feature.
}
