% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.getfasta.R
\name{bt.getfasta}
\alias{bt.getfasta}
\title{Extract DNA sequences from a fasta file based on feature coordinates.}
\usage{
bt.getfasta(
  fi,
  bed,
  fo = NULL,
  name = NULL,
  nameplus = NULL,
  nameOnly = NULL,
  split = NULL,
  tab = NULL,
  bedOut = NULL,
  s = NULL,
  fullHeader = NULL,
  rna = NULL,
  output = NULL
)
}
\arguments{
\item{fi}{<fasta>}

\item{bed}{<bed/gff/vcf>}

\item{fo}{Output file (opt., default is STDOUT}

\item{name}{Use the name field and coordinates for the FASTA header}

\item{nameplus}{(deprecated) Use the name field and coordinates for the FASTA header}

\item{nameOnly}{Use the name field for the FASTA header}

\item{split}{Given BED12 fmt., extract and concatenate the sequences
from the BED "blocks" (e.g., exons)}

\item{tab}{Write output in TAB delimited format.}

\item{bedOut}{Report extract sequences in a tab-delimited BED format instead of in FASTA format.
- Default is FASTA format.}

\item{s}{Force strandedness. If the feature occupies the antisense,
strand, the sequence will be reverse complemented.
- By default, strand information is ignored.}

\item{fullHeader}{Use full fasta header.
- By default, only the word before the first space or tab 
is used.}

\item{rna}{The FASTA is RNA not DNA. Reverse complementation handled accordingly.}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Extract DNA sequences from a fasta file based on feature coordinates.
}
