% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.groupby.R
\name{bt.groupby}
\alias{bt.groupby}
\title{Summarizes a dataset column based upon
common column groupings. Akin to the SQL "group by" command.}
\usage{
bt.groupby(
  i,
  g = NULL,
  c = NULL,
  o = NULL,
  full = NULL,
  inheader = NULL,
  outheader = NULL,
  header = NULL,
  ignorecase = NULL,
  prec = NULL,
  delim = NULL,
  output = NULL
)
}
\arguments{
\item{i}{<bed/gff/vcf/bam>}

\item{g}{<group columns>}

\item{c}{<op. column>}

\item{o}{<operation>}

\item{full}{Print all columns from input file.  The first line in the group is used.
Default: print only grouped columns.}

\item{inheader}{Input file has a header line - the first line will be ignored.}

\item{outheader}{Print header line in the output, detailing the column names. 
If the input file has headers (-inheader), the output file
will use the input's column names.
If the input file has no headers, the output file
will use "col_1", "col_2", etc. as the column names.}

\item{header}{same as '-inheader -outheader'}

\item{ignorecase}{Group values regardless of upper/lower case.}

\item{prec}{Sets the decimal precision for output (Default: 5)}

\item{delim}{Specify a custom delimiter for the collapse operations.
- Example: -delim "|"
- Default: ",".}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Summarizes a dataset column based upon
common column groupings. Akin to the SQL "group by" command.
}
