% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.links.R
\name{bt.links}
\alias{bt.links}
\title{Creates HTML links to an UCSC Genome Browser from a feature file.}
\usage{
bt.links(i, base = NULL, org = NULL, db = NULL, output = NULL)
}
\arguments{
\item{i}{<bed/gff/vcf> > out.html}

\item{base}{The browser basename.  Default: http://genome.ucsc.edu}

\item{org}{The organism. Default: human}

\item{db}{The build.  Default: hg18}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Creates HTML links to an UCSC Genome Browser from a feature file.
}
