% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.makewindows.R
\name{bt.makewindows}
\alias{bt.makewindows}
\title{Makes adjacent or sliding windows across a genome or BED file.}
\usage{
bt.makewindows(
  g = NULL,
  b = NULL,
  w = NULL,
  s = NULL,
  n = NULL,
  reverse = NULL,
  i = NULL,
  output = NULL
)
}
\arguments{
\item{g}{<genome> OR}

\item{b}{<bed>}

\item{w}{<window_size> OR}

\item{s}{<step_size>
Step size: i.e., how many base pairs to step before
creating a new window. Used to create "sliding" windows.
- Defaults to window size (non-sliding windows).}

\item{n}{<number of windows>}

\item{reverse}{Reverse numbering of windows in the output, i.e. report 
windows in decreasing order}

\item{i}{src|winnum|srcwinnum
The default output is 3 columns: chrom, start, end .
With this option, a name column will be added.
 "-i src" - use the source interval's name.
 "-i winnum" - use the window number as the ID (e.g. 1,2,3,4...).
 "-i srcwinnum" - use the source interval's name with the window number.
See below for usage examples.}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Makes adjacent or sliding windows across a genome or BED file.
}
