% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.nuc.R
\name{bt.nuc}
\alias{bt.nuc}
\title{Profiles the nucleotide content of intervals in a fasta file.}
\usage{
bt.nuc(
  fi,
  bed,
  s = NULL,
  seq = NULL,
  pattern = NULL,
  C = NULL,
  fullHeader = NULL,
  output = NULL
)
}
\arguments{
\item{fi}{<fasta>}

\item{bed}{<bed/gff/vcf>}

\item{s}{Profile the sequence according to strand.}

\item{seq}{Print the extracted sequence}

\item{pattern}{Report the number of times a user-defined sequence
is observed (case-sensitive).}

\item{C}{Ignore case when matching -pattern. By defaulty, case matters.}

\item{fullHeader}{Use full fasta header.
- By default, only the word before the first space or tab is used.}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Profiles the nucleotide content of intervals in a fasta file.
}
