% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.pairtopair.R
\name{bt.pairtopair}
\alias{bt.pairtopair}
\title{Report overlaps between two paired-end BED files (BEDPE).}
\usage{
bt.pairtopair(a, b, f = NULL, type = NULL, slop = NULL, ss = NULL,
  is = NULL, rdn = NULL, output = NULL)
}
\arguments{
\item{a}{<BEDPE>}

\item{b}{<BEDPE>}

\item{f}{Minimum overlap required as fraction of A (e.g. 0.05).
Default is 1E-9 (effectively 1bp).}

\item{type}{Approach to reporting overlaps between A and B.
neither Report overlaps if neither end of A overlaps B.
either Report overlaps if either ends of A overlap B.
both Report overlaps if both ends of A overlap B.
notboth Report overlaps if one or neither of A's overlap B.
- Default = both.}

\item{slop}{The amount of slop (in b.p.). to be added to each footprint of A.
*Note*: Slop is subtracted from start1 and start2
  and added to end1 and end2.
- Default = 0.}

\item{ss}{Add slop based to each BEDPE footprint based on strand.
- If strand is "+", slop is only added to the end coordinates.
- If strand is "-", slop is only added to the start coordinates.
- By default, slop is added in both directions.}

\item{is}{Ignore strands when searching for overlaps.
- By default, strands are enforced.}

\item{rdn}{Require the hits to have different names (i.e. avoid self-hits).
- By default, same names are allowed.}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Report overlaps between two paired-end BED files (BEDPE).
}
