% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.random.R
\name{bt.random}
\alias{bt.random}
\title{Generate random intervals among a genome.}
\usage{
bt.random(g, l = NULL, n = NULL, seed = NULL, output = NULL)
}
\arguments{
\item{g}{<genome>}

\item{l}{The length of the intervals to generate.
- Default = 100.
- (INTEGER)}

\item{n}{The number of intervals to generate.
- Default = 1,000,000.
- (INTEGER)}

\item{seed}{Supply an integer seed for the shuffling.
- By default, the seed is chosen automatically.
- (INTEGER)}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Generate random intervals among a genome.
}
