% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.reldist.R
\name{bt.reldist}
\alias{bt.reldist}
\title{Calculate the relative distance distribution b/w two feature files.}
\usage{
bt.reldist(a, b, detail = NULL, output = NULL)
}
\arguments{
\item{a}{<bed/gff/vcf>}

\item{b}{<bed/gff/vcf>}

\item{detail}{Report the relativedistance for each interval in A}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Calculate the relative distance distribution b/w two feature files.
}
