% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.sample.R
\name{bt.sample}
\alias{bt.sample}
\title{Take sample of input file(s) using reservoir sampling algorithm.}
\usage{
bt.sample(
  i,
  n = NULL,
  seed = NULL,
  ubam = NULL,
  s = NULL,
  header = NULL,
  bed = NULL,
  nobuf = NULL,
  iobuf = NULL,
  output = NULL
)
}
\arguments{
\item{i}{<bed/gff/vcf/bam>}

\item{n}{The number of records to generate.
- Default = 1,000,000.
- (INTEGER)}

\item{seed}{Supply an integer seed for the shuffling.
- By default, the seed is chosen automatically.
- (INTEGER)}

\item{ubam}{Write uncompressed BAM output. Default writes compressed BAM.}

\item{s}{Require same strandedness.  That is, only give records
that have the same strand. Use '-s forward' or '-s reverse'
for forward or reverse strand records, respectively.
- By default, records are reported without respect to strand.}

\item{header}{Print the header from the A file prior to results.}

\item{bed}{If using BAM input, write output as BED.}

\item{nobuf}{Disable buffered output. Using this option will cause each line
of output to be printed as it is generated, rather than saved
in a buffer. This will make printing large output files 
noticeably slower, but can be useful in conjunction with
other software tools and scripts that need to process one
line of bedtools output at a time.}

\item{iobuf}{Specify amount of memory to use for input buffer.
Takes an integer argument. Optional suffixes K/M/G supported.
Note: currently has no effect with compressed files.}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Take sample of input file(s) using reservoir sampling algorithm.
}
