% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.sort.R
\name{bt.sort}
\alias{bt.sort}
\title{Sorts a feature file in various and useful ways.}
\usage{
bt.sort(
  i,
  sizeA = NULL,
  sizeD = NULL,
  chrThenSizeA = NULL,
  chrThenSizeD = NULL,
  chrThenScoreA = NULL,
  chrThenScoreD = NULL,
  g = NULL,
  faidx = NULL,
  header = NULL,
  output = NULL
)
}
\arguments{
\item{i}{<bed/gff/vcf>}

\item{sizeA}{Sort by feature size in ascending order.}

\item{sizeD}{Sort by feature size in descending order.}

\item{chrThenSizeA}{Sort by chrom (asc), then feature size (asc).}

\item{chrThenSizeD}{Sort by chrom (asc), then feature size (desc).}

\item{chrThenScoreA}{Sort by chrom (asc), then score (asc).}

\item{chrThenScoreD}{Sort by chrom (asc), then score (desc).}

\item{g}{Sort according to the chromosomes declared in "genome.txt"}

\item{faidx}{Sort according to the chromosomes declared in "names.txt"}

\item{header}{Print the header from the A file prior to results.}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Sorts a feature file in various and useful ways.
}
