% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.spacing.R
\name{bt.spacing}
\alias{bt.spacing}
\title{Report (last col.) the gap lengths between intervals in a file.}
\usage{
bt.spacing(i, output = NULL)
}
\arguments{
\item{i}{<bed/gff/vcf/bam>}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Report (last col.) the gap lengths between intervals in a file.
}
