% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.split.R
\name{bt.split}
\alias{bt.split}
\title{Split a Bed file.}
\usage{
bt.split(i, n = NULL, p = NULL, a = NULL)
}
\arguments{
\item{i}{<bed>}

\item{n}{number-of-files}

\item{p}{Output BED file prefix.}

\item{a}{Algorithm used to split data.
* size (default): uses a heuristic algorithm to group the items
  so all files contain the ~ same number of bases
* simple : route records such that each split file has
  approximately equal records (like Unix split).}
}
\description{
Split a Bed file.
}
