% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.summary.R
\name{bt.summary}
\alias{bt.summary}
\title{Report summary statistics of the intervals in a file}
\usage{
bt.summary(i, g, output = NULL)
}
\arguments{
\item{i}{<bed/gff/vcf/bam>}

\item{g}{<genome>}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Report summary statistics of the intervals in a file
}
