% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Clusters overlapping/nearby BED/GFF/VCF intervals.}
\usage{
cluster(i, s = NULL, d = NULL, output = NULL)
}
\arguments{
\item{i}{<bed/gff/vcf>}

\item{s}{Force strandedness.  That is, only merge features
that are the same strand.
- By default, merging is done without respect to strand.}

\item{d}{Maximum distance between features allowed for features
to be merged.
- Def. 0. That is, overlapping & book-ended features are merged.
- (INTEGER)}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Clusters overlapping/nearby BED/GFF/VCF intervals.
}
