% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOptions.r
\name{createOptions}
\alias{createOptions}
\title{Creates options based on user input}
\usage{
createOptions(names, values)
}
\arguments{
\item{names}{vector of names of options}

\item{values}{vector of values of options

### Define a function that determines establishes files and paths for bedtools functions}
}
\description{
Creates options based on user input
}
