% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteTempFiles.R
\name{deleteTempFiles}
\alias{deleteTempFiles}
\title{Deletes temp files}
\usage{
deleteTempFiles(tempfiles)
}
\arguments{
\item{tempfiles}{a vector of tempfiles for deletion

### Define a function that determines establishes files and paths for bedtools functions}
}
\description{
Deletes temp files
}
