% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establishPaths.R
\name{establishPaths}
\alias{establishPaths}
\title{Determines if arguments are paths or R objects. Makes temp files when
neccesary. Makes a list of files to use in bedtools call. Makes a list
of temp files to delete at end of function}
\usage{
establishPaths(input, name = "", allowRobjects = TRUE)
}
\arguments{
\item{input}{the input for an argument.  Could be a path to a file, an R object (data frame), or a list of any combination thereof}

\item{name}{the name of the argument}

\item{allowRobjects}{boolean whether or not to allow R objects as inputs

### Define a function that determines establishes files and paths for bedtools functions}
}
\description{
Determines if arguments are paths or R objects. Makes temp files when
neccesary. Makes a list of files to use in bedtools call. Makes a list
of temp files to delete at end of function
}
