% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Replicate lines in a file based on columns of comma-separated values.}
\usage{
expand(c, i = NULL)
}
\arguments{
\item{c}{}

\item{i}{Input file. Assumes "stdin" if omitted.}
}
\description{
Replicate lines in a file based on columns of comma-separated values.
}
