% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.r
\name{fisher}
\alias{fisher}
\title{Calculate Fisher statistic b/w two feature files.}
\usage{
fisher(a, b, g, m = NULL, s = NULL, S = NULL, f = NULL, F = NULL,
  r = NULL, e = NULL, split = NULL, nonamecheck = NULL,
  bed = NULL, header = NULL, nobuf = NULL, iobuf = NULL)
}
\arguments{
\item{a}{<bed/gff/vcf>}

\item{b}{<bed/gff/vcf>}

\item{g}{<genome file>}

\item{m}{Merge overlapping intervals before
- looking at overlap.}

\item{s}{Require same strandedness.  That is, only report hits in B
that overlap A on the _same_ strand.
- By default, overlaps are reported without respect to strand.}

\item{S}{Require different strandedness.  That is, only report hits in B
that overlap A on the _opposite_ strand.
- By default, overlaps are reported without respect to strand.}

\item{f}{Minimum overlap required as a fraction of A.
- Default is 1E-9 (i.e., 1bp).
- FLOAT (e.g. 0.50)}

\item{F}{Minimum overlap required as a fraction of B.
- Default is 1E-9 (i.e., 1bp).
- FLOAT (e.g. 0.50)}

\item{r}{Require that the fraction overlap be reciprocal for A AND B.
- In other words, if -f is 0.90 and -r is used, this requires
  that B overlap 90 percent of A and A _also_ overlaps 90 percent of B.}

\item{e}{Require that the minimum fraction be satisfied for A OR B.
- In other words, if -e is used with -f 0.90 and -F 0.10 this requires
  that either 90 percent of A is covered OR 10 percent of  B is covered.
  Without -e, both fractions would have to be satisfied.}

\item{split}{Treat "split" BAM or BED12 entries as distinct BED intervals.}

\item{nonamecheck}{For sorted data, don't throw an error if the file has different naming conventions
for the same chromosome. ex. "chr1" vs "chr01".}

\item{bed}{If using BAM input, write output as BED.}

\item{header}{Print the header from the A file prior to results.}

\item{nobuf}{Disable buffered output. Using this option will cause each line
of output to be printed as it is generated, rather than saved
in a buffer. This will make printing large output files 
noticeably slower, but can be useful in conjunction with
other software tools and scripts that need to process one
line of bedtools output at a time.}

\item{iobuf}{Specify amount of memory to use for input buffer.
Takes an integer argument. Optional suffixes K/M/G supported.
Note: currently has no effect with compressed files.}
}
\description{
Calculate Fisher statistic b/w two feature files.
}
