% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfasta.r
\name{getfasta}
\alias{getfasta}
\title{Extract DNA sequences from a fasta file based on feature coordinates.}
\usage{
getfasta(fi, bed, name = NULL, fullHeader = NULL, nameplus = NULL,
  s = NULL, split = NULL, tab = NULL, fo = NULL)
}
\arguments{
\item{fi}{<fasta>}

\item{bed}{<bed/gff/vcf>}

\item{name}{Use the name field for the FASTA header}

\item{fullHeader}{Use full fasta header.
- By default, only the word before the first space or tab 
is used.}

\item{nameplus}{Use the name field and coordinates for the FASTA header}

\item{s}{Force strandedness. If the feature occupies the antisense,
strand, the sequence will be reverse complemented.
- By default, strand information is ignored.}

\item{split}{given BED12 fmt., extract and concatenate the sequences
from the BED "blocks" (e.g., exons)}

\item{tab}{Write output in TAB delimited format.
- Default is FASTA format.}

\item{fo}{Output file (opt., default is STDOUT}
}
\description{
Extract DNA sequences from a fasta file based on feature coordinates.
}
