% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.r
\name{random}
\alias{random}
\title{Generate random intervals among a genome.}
\usage{
random(g, l = NULL, n = NULL, seed = NULL)
}
\arguments{
\item{g}{<genome>}

\item{l}{The length of the intervals to generate.
- Default = 100.
- (INTEGER)}

\item{n}{The number of intervals to generate.
- Default = 1,000,000.
- (INTEGER)}

\item{seed}{Supply an integer seed for the shuffling.
- By default, the seed is chosen automatically.
- (INTEGER)}
}
\description{
Generate random intervals among a genome.
}
