% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.r
\name{tag}
\alias{tag}
\title{Annotates a BAM file based on overlaps with multiple BED/GFF/VCF files
on the intervals in -i.}
\usage{
tag(i, files, labels, f = NULL, S = NULL, intervals = NULL,
  s = NULL, tag = NULL, names = NULL, scores = NULL)
}
\arguments{
\item{i}{<BAM>}

\item{files}{FILE1 .. FILEn}

\item{labels}{LAB1 .. LABn}

\item{f}{Minimum overlap required as a fraction of the alignment.
- Default is 1E-9 (i.e., 1bp).
- FLOAT (e.g. 0.50)}

\item{S}{Require overlaps on the opposite strand.  That is, only tag alignments that have the opposite
strand as a feature in the annotation file(s).}

\item{intervals}{Use the full interval (including name, score, and strand) to populate tags.
Requires the -labels option to identify from which file the interval came.}

\item{s}{Require overlaps on the same strand.  That is, only tag alignments that have the same
strand as a feature in the annotation file(s).}

\item{tag}{Dictate what the tag should be. Default is YB.
- STRING (two characters, e.g., YK)}

\item{names}{Use the name field from the annotation files to populate tags.
By default, the -labels values are used.}

\item{scores}{Use the score field from the annotation files to populate tags.
By default, the -labels values are used.}
}
\description{
Annotates a BAM file based on overlaps with multiple BED/GFF/VCF files
on the intervals in -i.
}
