% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedtobam.r
\name{bedtobam}
\alias{bedtobam}
\title{Converts feature records to BAM format.}
\usage{
bedtobam(i, g, ubam = NULL, mapq = NULL, bed12 = NULL)
}
\arguments{
\item{i}{<bed/gff/vcf>}

\item{g}{<genome>}

\item{ubam}{Write uncompressed BAM output. Default writes compressed BAM.}

\item{mapq}{Set the mappinq quality for the BAM records.
(INT) Default: 255}

\item{bed12}{The BED file is in BED12 format.  The BAM CIGAR
string will reflect BED "blocks".}
}
\description{
Converts feature records to BAM format.
}
