% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.expand.R
\name{bt.expand}
\alias{bt.expand}
\title{Replicate lines in a file based on columns of comma-separated values.}
\usage{
bt.expand(i, c = NULL, output = NULL)
}
\arguments{
\item{i}{<bed/gff/vcf/bam>}

\item{c}{Specify the column (1-based) that should be summarized.
- Required.}

\item{output}{Output filepath instead of returning output in R.}
}
\description{
Replicate lines in a file based on columns of comma-separated values.
}
