% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteTempfiles.r
\name{deleteTempfiles}
\alias{deleteTempfiles}
\title{Deletes temp files}
\usage{
deleteTempfiles(tempfiles)
}
\arguments{
\item{tempfiles}{a vector of tempfiles for deletion

### Define a function that determines establishes files and paths for bedtools functions}
}
\description{
Deletes temp files
}
