% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand.R
\name{expand}
\alias{expand}
\title{Replicate lines in a file based on columns of comma-separated values.}
\usage{
expand(i, c = NULL)
}
\arguments{
\item{i}{<bed/gff/vcf/bam>}

\item{c}{}
}
\description{
Replicate lines in a file based on columns of comma-separated values.
}
