% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{Computes the amount of overlap (positive values)
or distance (negative values) between genome features
and reports the result at the end of the same line.}
\usage{
overlap(i, cols = NULL)
}
\arguments{
\item{i}{<bed/gff/vcf/bam>}

\item{cols}{<columns>}
}
\description{
Computes the amount of overlap (positive values)
or distance (negative values) between genome features
and reports the result at the end of the same line.
}
